[frontend] new way to refacto widgets
